% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{check_for_updates}
\alias{check_for_updates}
\title{Title check_for_updates}
\usage{
check_for_updates(repo = "")
}
\arguments{
\item{repo}{}
}
\description{
Title check_for_updates
}
\examples{
\donttest{
rdataretriever::check_for_updates()
}
}
