% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{check_retriever_availability}
\alias{check_retriever_availability}
\title{Check to see if minimum version of retriever Python package is installed}
\usage{
check_retriever_availability()
}
\value{
boolean
}
\description{
Check to see if minimum version of retriever Python package is installed
}
\examples{
\dontrun{
rdataretriever::check_retriever_availability()
}
}
