% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{commit}
\alias{commit}
\title{Title commit}
\usage{
commit(dataset, commit_message = "", path = NULL, quiet = FALSE)
}
\arguments{
\item{dataset}{}

\item{commit_message}{}

\item{path}{}

\item{quiet}{}
}
\description{
Title commit
}
\examples{
\donttest{
rdataretriever::commit('iris')
}
}
