% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{commit_log}
\alias{commit_log}
\title{Title commit_log}
\usage{
commit_log(dataset)
}
\arguments{
\item{dataset}{}
}
\description{
Title commit_log
}
\examples{
\donttest{
rdataretriever::commit_log('iris')
}
}
