% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{create_rdataset}
\alias{create_rdataset}
\title{Download files for RDatasets to the raw data directory}
\usage{
create_rdataset(engine, package, dataset_name, script_path = NULL)
}
\arguments{
\item{engine}{}

\item{package}{}

\item{dataset_name}{}

\item{script_path}{}
}
\description{
Download files for RDatasets to the raw data directory
}
\examples{
\donttest{
rdataretriever::create_rdataset(engine, package, dataset_name)
}
}
