% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{create_socrata_dataset}
\alias{create_socrata_dataset}
\title{Downloads raw data and creates a script for the socrata dataset}
\usage{
create_socrata_dataset(engine, name, resource, script_path = NULL)
}
\arguments{
\item{engine}{}

\item{name}{}

\item{resource}{}

\item{script_path}{}
}
\description{
Downloads raw data and creates a script for the socrata dataset
}
\examples{
\donttest{
rdataretriever::create_socrata_dataset(engine, name, resource)
}
}
