% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{data_retriever_version}
\alias{data_retriever_version}
\title{Get Data Retriever version}
\usage{
data_retriever_version(clean = TRUE)
}
\arguments{
\item{clean}{boolean return cleaned version appropriate for semver}
}
\value{
returns a string with the version information
}
\description{
Get Data Retriever version
}
\examples{
\dontrun{
rdataretriever::data_retriever_version()
}
}
