% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{dataset_names}
\alias{dataset_names}
\title{Name all available dataset scripts.}
\usage{
dataset_names()
}
\value{
returns a character vector with the available datasets for download
}
\description{
Additional information on the available datasets can be found at url https://retriever.readthedocs.io/en/latest/datasets.html
}
\examples{
\dontrun{
rdataretriever::dataset_names()
}
}
