% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{datasets}
\alias{datasets}
\title{Name all available dataset scripts.}
\usage{
datasets(keywords = "", licenses = "")
}
\arguments{
\item{keywords}{search all datasets by keywords}

\item{licenses}{search all datasets by licenses}
}
\value{
returns a character vector with the available datasets for download

Returns the names of all available dataset scripts
}
\description{
Additional information on the available datasets can be found at url https://retriever.readthedocs.io/en/latest/datasets.html
}
\examples{
\dontrun{
rdataretriever::datasets()
}
}
