% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{display_all_rdataset_names}
\alias{display_all_rdataset_names}
\title{Displays the list of rdataset names present in the packages provided}
\usage{
display_all_rdataset_names(package_name = NULL)
}
\arguments{
\item{package_name}{}
}
\description{
Displays the list of rdataset names present in the packages provided
}
\examples{
\donttest{
rdataretriever::display_all_rdataset_names()
}
}
