% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{find_socrata_dataset_by_id}
\alias{find_socrata_dataset_by_id}
\title{Returns metadata for the following dataset id}
\usage{
find_socrata_dataset_by_id(dataset_id)
}
\arguments{
\item{dataset_id}{}
}
\description{
Returns metadata for the following dataset id
}
\examples{
\donttest{
rdataretriever::socrata_dataset_info()
}
}
