% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{get_citation}
\alias{get_citation}
\title{Get dataset citation information and a description}
\usage{
get_citation(dataset)
}
\value{
returns a character vector with the available datasets for download
}
\description{
Get dataset citation information and a description
}
\examples{
\donttest{
rdataretriever::get_citation('plant-comp-us')
}
}
