% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{get_dataset_names_upstream}
\alias{get_dataset_names_upstream}
\title{Title}
\usage{
get_dataset_names_upstream(keywords = "", licenses = "", repo = "")
}
\arguments{
\item{keywords}{}

\item{licenses}{}

\item{repo}{}
}
\description{
Title
}
