% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{get_rdataset_names}
\alias{get_rdataset_names}
\title{Returns a list of all the available RDataset names present}
\usage{
get_rdataset_names()
}
\value{
No return value, list of all the available RDataset
}
\description{
Returns a list of all the available RDataset names present
}
\examples{
\dontrun{
rdataretriever::get_rdataset_names()
}
}
