% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{get_script_citation}
\alias{get_script_citation}
\title{Get citation}
\usage{
get_script_citation(dataset = "")
}
\arguments{
\item{dataset}{dataset to obtain citation}
}
\description{
Get citation
}
\examples{
\donttest{
rdataretriever::get_script_citation(dataset = "")
}
}
