% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{get_version}
\alias{get_version}
\title{Get retriever version}
\usage{
get_version()
}
\value{
returns a string with the version information
}
\description{
Get retriever version
}
\examples{
\donttest{
rdataretriever::get_version()
}
}
