% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{install_mysql}
\alias{install_mysql}
\title{Title}
\usage{
install_mysql(dataset, user = "root", password = "", host = "localhost",
  port = 3306, database_name = "{db}", table_name = "{db}.{table}",
  debug = FALSE, use_cache = TRUE)
}
\arguments{
\item{dataset}{}

\item{user}{}

\item{password}{}

\item{host}{}

\item{port}{}

\item{database_name}{}

\item{table_name}{}

\item{debug}{}

\item{use_cache}{}
}
\description{
Title
}
