% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{install_postgres}
\alias{install_postgres}
\title{Title}
\usage{
install_postgres(dataset, user = "postgres", password = "",
  host = "localhost", port = 5432, database = "postgres",
  database_name = "{db}", table_name = "{db}.{table}", debug = FALSE,
  use_cache = TRUE)
}
\arguments{
\item{dataset}{}

\item{user}{}

\item{password}{}

\item{host}{}

\item{port}{}

\item{database}{}

\item{database_name}{}

\item{table_name}{}

\item{debug}{}

\item{use_cache}{}
}
\description{
Title
}
