% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{install_sqlite}
\alias{install_sqlite}
\title{Title}
\usage{
install_sqlite(dataset, file = "sqlite.db", table_name = "{db}_{table}",
  debug = FALSE, use_cache = TRUE)
}
\arguments{
\item{dataset}{}

\item{file}{}

\item{table_name}{}

\item{debug}{}

\item{use_cache}{}
}
\description{
Title
}
