% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{install_xml}
\alias{install_xml}
\title{Title}
\usage{
install_xml(dataset, table_name = "{db}_{table}.xml", debug = FALSE,
  use_cache = TRUE)
}
\arguments{
\item{dataset}{}

\item{table_name}{}

\item{debug}{}

\item{use_cache}{}
}
\description{
Title
}
