% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{reload_scripts}
\alias{reload_scripts}
\title{Update the retriever's global_script_list with the scripts present
in the ~/.retriever directory.}
\usage{
reload_scripts()
}
\value{
No return value, fetches most resent scripts
}
\description{
Update the retriever's global_script_list with the scripts present
in the ~/.retriever directory.
}
\examples{
\dontrun{
rdataretriever::reload_scripts()
}
}
