% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{reset}
\alias{reset}
\title{Reset the scripts or data(raw_data) directory or both}
\usage{
reset(scope = "all")
}
\arguments{
\item{scope}{All resets both scripst and data directory}
}
\value{
No return value, resets the scripts and the data directory
}
\description{
Reset the scripts or data(raw_data) directory or both
}
\examples{
\dontrun{
rdataretriever::reset("iris")
}
}
