% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{retriever_meets_min_version}
\alias{retriever_meets_min_version}
\title{Check Data Retriever version satisfies requirements}
\usage{
retriever_meets_min_version()
}
\value{
boolean
}
\description{
Check Data Retriever version satisfies requirements
}
\examples{
\donttest{
rdataretriever::check_data_retriever_version()
}
}
