% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{socrata_dataset_info}
\alias{socrata_dataset_info}
\title{Title}
\usage{
socrata_dataset_info(dataset_name)
}
\arguments{
\item{dataset_name}{}
}
\description{
Title
}
\examples{
\donttest{
rdataretriever::socrata_dataset_info()
}
}
