% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{update_rdataset_catalog}
\alias{update_rdataset_catalog}
\title{Updates the datasets_url.json from the github repo}
\usage{
update_rdataset_catalog(test = FALSE)
}
\arguments{
\item{test}{flag set when testing}
}
\value{
No return value, updates the datasets_url.json
}
\description{
Updates the datasets_url.json from the github repo
}
\examples{
\dontrun{
rdataretriever::update_rdataset_catalog()
}
}
