% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{update_rdataset_contents}
\alias{update_rdataset_contents}
\title{Update the contents of json script}
\usage{
update_rdataset_contents(data_obj, package, dataset_name, json_file)
}
\arguments{
\item{data_obj}{}

\item{package}{}

\item{dataset_name}{}

\item{json_file}{}
}
\description{
Update the contents of json script
}
\examples{
\donttest{
rdataretriever::update_rdataset_contents(data_obj, package, dataset_name, json_file)
}
}
