% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{update_rdataset_script}
\alias{update_rdataset_script}
\title{Renames and updates the RDataset script}
\usage{
update_rdataset_script(data_obj, dataset_name, package, script_path)
}
\arguments{
\item{data_obj}{}

\item{dataset_name}{}

\item{package}{}

\item{script_path}{}
}
\description{
Renames and updates the RDataset script
}
\examples{
\donttest{
rdataretriever::update_rdataset_script(data_obj, dataset_name, package, script_path)
}
}
