% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{update_socrata_contents}
\alias{update_socrata_contents}
\title{Update the contents of the json script}
\usage{
update_socrata_contents(json_file, script_name, url, resource)
}
\arguments{
\item{json_file}{}

\item{script_name}{}

\item{url}{}

\item{resource}{}
}
\description{
Update the contents of the json script
}
\examples{
\donttest{
rdataretriever::update_socrata_contents(json_file, script_name, url, resource)
}
}
