% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{update_socrata_script}
\alias{update_socrata_script}
\title{Renames the script name and the contents of the script}
\usage{
update_socrata_script(script_name, filename, url, resource, script_path)
}
\arguments{
\item{script_name}{}

\item{filename}{}

\item{url}{}

\item{resource}{}

\item{script_path}{}
}
\description{
Renames the script name and the contents of the script
}
\examples{
\donttest{
rdataretriever::update_socrata_script(script_name, filename, url, resource, script_path)
}
}
