% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{use_RetrieverPath}
\alias{use_RetrieverPath}
\title{Setting path of retriever}
\usage{
use_RetrieverPath(path)
}
\arguments{
\item{path}{location of retriever in the system}
}
\value{
No return value, setting path of retriever
}
\description{
Setting path of retriever
}
\examples{
\dontrun{
rdataretriever::use_RetrieverPath("/home/<system_name>/anaconda2/envs/py27/bin/")
}
}
