% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{check_for_updates}
\alias{check_for_updates}
\title{Check for updates}
\usage{
check_for_updates(repo = "")
}
\arguments{
\item{repo}{path to the repository}
}
\description{
Check for updates
}
\examples{
\donttest{
rdataretriever::check_for_updates()
}
}
