% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{commit}
\alias{commit}
\title{Commit a dataset}
\usage{
commit(dataset, commit_message = "", path = NULL, quiet = FALSE)
}
\arguments{
\item{dataset}{name of the dataset}

\item{commit_message}{commit message for the commit}

\item{path}{path to save the committed dataset, if no path given save in provenance directory}

\item{quiet}{logical, if true retriever runs in quiet mode}
}
\value{
No return value, provides confirmation for commit
}
\description{
Commit a dataset
}
\examples{
\dontrun{
rdataretriever::commit("iris")
}
}
