% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{commit_log}
\alias{commit_log}
\title{See the log of committed dataset stored in provenance directory}
\usage{
commit_log(dataset)
}
\arguments{
\item{dataset}{name of the dataset stored in provenance directory}
}
\value{
No return value, prints message after commit
}
\description{
See the log of committed dataset stored in provenance directory
}
\examples{
\dontrun{
rdataretriever::commit_log("iris")
}
}
