% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{download}
\alias{download}
\title{Title}
\usage{
download(dataset, path = "./", quiet = FALSE, sub_dir = FALSE,
  debug = FALSE)
}
\arguments{
\item{dataset}{}

\item{path}{}

\item{quiet}{}

\item{sub_dir}{}

\item{debug}{}
}
\description{
Title
}
