% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{get_retriever_citation}
\alias{get_retriever_citation}
\title{Get retriever citation}
\usage{
get_retriever_citation()
}
\value{
No return value, outputs citation of the Data Retriever Python package
}
\description{
Get retriever citation
}
\examples{
\dontrun{
rdataretriever::get_retriever_citation()
}
}
