% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{get_script_citation}
\alias{get_script_citation}
\title{Get citation of a script}
\usage{
get_script_citation(dataset = NULL)
}
\arguments{
\item{dataset}{dataset to obtain citation}
}
\value{
No return value, gets citation of a script
}
\description{
Get citation of a script
}
\examples{
\dontrun{
rdataretriever::get_script_citation(dataset = "")
}
}
