% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{get_script_upstream}
\alias{get_script_upstream}
\title{Get script upstream}
\usage{
get_script_upstream(dataset, repo = "")
}
\arguments{
\item{dataset}{name of the dataset}

\item{repo}{path to the repository}
}
\description{
Get script upstream
}
\examples{
\donttest{
rdataretriever::get_script_upstream("iris")
}
}
