% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{socrata_autocomplete_search}
\alias{socrata_autocomplete_search}
\title{Returns the list of dataset names after autocompletion}
\usage{
socrata_autocomplete_search(dataset)
}
\arguments{
\item{dataset}{the name of the dataset}
}
\value{
No return value, show dataset names after autocompletion
}
\description{
Returns the list of dataset names after autocompletion
}
\examples{
\dontrun{
rdataretriever::socrata_autocomplete_search()
}
}
