% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{socrata_dataset_info}
\alias{socrata_dataset_info}
\title{Get socrata dataset info}
\usage{
socrata_dataset_info(dataset_name)
}
\arguments{
\item{dataset_name}{dataset name to obtain info}
}
\value{
No return value, shows socrata dataset info
}
\description{
Get socrata dataset info
}
\examples{
\dontrun{
rdataretriever::socrata_dataset_info()
}
}
